add set/get gyro filter, and allow uint16 function https://bitbucket.org/lpresearch/lpmsig1firmware/src/297e5e7ac726ae89ad6cf35c205bc787edfe2860/Program/LPMS-IG1/Inc/LpmsConfig.h#lines-268

+++ src/Sensor.cpp
@@ -331,7 +331,7 @@ namespace zen
             SensorManager::get().release({ m_token });
     }
 
-    ZenError Sensor::processReceivedData(uint8_t, uint8_t function, gsl::span<const std::byte> data) noexcept
+    ZenError Sensor::processReceivedData(uint8_t, uint16_t function, gsl::span<const std::byte> data) noexcept
     {
         if (m_config.version == 0)
         {
